requirements.txt
absl-py==0.3.0
astor==0.7.1
astroid==2.0.1
bleach==1.5.0
Brotli==1.0.0
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
click==6.7
cloudpickle==0.4.1
colorama==0.3.9
cycler==0.10.0
Cython==0.28.4
dask==0.15.4
decorator==4.2.1
Django==2.0.7
dlib==19.8.1
docopt==0.6.2
face-alignment==0.1.0
face-recognition==1.2.1
face-recognition-models==0.3.0
facenet==1.0.3
Flask==1.0.2
future==0.16.0
gast==0.2.0
gevent==1.3.5
greenlet==0.4.14
grpcio==1.14.0
h5py==2.7.1
html5lib==0.9999999
idna==2.7
image==1.5.24
imutils==0.4.6
isort==4.3.4
itsdangerous==0.24
jedi==0.11.0
Jinja2==2.10
joblib==0.11
Keras==2.1.2
keras-contrib==2.0.8
labelme==3.3.3
lazy-object-proxy==1.3.1
locket==0.2.0
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.1.2
mccabe==0.6.1
msgpack-python==0.5.6
networkx==2.1
numpy==1.15.0
opencv-python==3.4.2.16
parso==0.1.0
partd==0.3.8
pathlib==1.0.1
Pillow==5.0.0
prompt-toolkit==1.0.15
protobuf==3.6.0
psutil==5.4.6
ptpython==0.41
pycparser==2.18
Pygments==2.2.0
PyJWT==1.6.4
pylint==2.0.1
pyodbc==4.0.19
pyparsing==2.2.0
pypiwin32==223
PyQt5==5.11.2
PyQt5-sip==4.19.12
pyreadline==2.0
PySocks==1.6.8
python-dateutil==2.6.1
python-snappy==0.5.1
pythonnet==2.4.0.dev0
pytz==2018.3
PyWavelets==0.5.2
pywin32==223
PyYAML==3.12
pyzmq==17.1.0
QtPy==1.4.2
requests==2.19.1
scandir==1.6
scikit-image==0.13.1
scikit-learn==0.19.2
scipy==1.0.0
sip==4.19.8
six==1.11.0
skflow==0.1.0
sklearn==0.0
sqlite-bro==0.8.11
tensorboard==1.9.0
tensorflow-gpu==1.9.0
termcolor==1.1.0
toolz==0.8.2
tqdm==4.19.6
twilio==6.16.1
typed-ast==1.1.0
urllib3==1.23
wcwidth==0.1.7
Werkzeug==0.14.1
winpython==1.9.20171031
wrapt==1.10.11
zerorpc==0.6.1
